Modified the make methods to reuse type information.
The make descriptor methods were repeating the type information that
was given in the type aliases for the descriptor types.  Now, the
type parameters given in the type aliases can be retrieved from the
instantiation using a type method as suggested by Brad:

https://stackoverflow.com/a/49951164/594274

The make descriptor methods are now only a convenience, and one could
just directly call the Wrapper's make method by referring to the type
aliases in AdjListHyperGraph and calling make on them directly (exactly
as is done in the convenience methods).