For the butterfly test, ake a graph argument instead
As per Louis, verticesDomain is accessing the field directly via `this`.verticesDomain where `this` is remote.
`this` is the class instance of the original locale that invoked the method
And when the algo jump to a remote locale, accessing this explicitly or implicitly results in communication.
CylicDist  is optimized to disable locality checks when accessing the DSI methods directly, like we are doing
We use graph.verticesDomain instead it will resolve it through privatization (I.E through record wrapper forwarding).